+++ lizard_auth_client/apps.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.apps import AppConfig
+
+
+class LizardAuthClientConfig(AppConfig):
+    name = 'lizard_auth_client'
+    verbose_name = "Lizard SSO authentication client"